Update README.md
This project aims to develop a machine learning model to predict the likelihood of stroke based on patient data. The primary objective is to identify all potential stroke cases, with a strong focus on maximizing sensitivity (recall) to ensure no stroke cases are missed. This project uses real-world-inspired healthcare data, making it applicable for practical use in early diagnosis and prevention strategies.